requirements.txt
pywebview==4.4.1
websockets==12.0
pyinstaller==6.3.0
